Add include directive. internal::Function now copies function body AST.
&include is now a supported directive to use to include files. Many files may
be included per include directive. Cyclic includes are not detected yet, but
I want to detect them and not have to worry about them.

Before, internal::Function objects would hold a Vec of AST references (so the
vector was owned by the function, but the AST objects were not). With the
introduction of include files where AST items can be coming from this multitude
of places, I found it easier just to make a copy of the AST. Unused RAM is
wasted RAM, right?!