Add error chaining.
I've been wanting to get this implemented for a while. We've reached the
critical complexity point where if it was put off any longer, it would be days
(compared to hours) of work to get this thing refactored so it's using the
error chaining nicely. A lot of this code looks like garbage because it's
basically replacing the hacky "error chaining" that was printing errors before.

There are some errors that aren't so pretty (such as errors with line numbers)
but that will have to come with time. I also want to start returning errors with
more definitive ranges in the source code so I can start printing out snippets
of the source where things went wrong.