Fix some dangling warnings, prune AST after preprocessing.
AST pruning was one of the bigger things that I was worrying about after
preprocessing. I was planning on having the gatherer remove anything that it
was attempting to preprocess but the way things are structured, I couldn't get
that method to work. So instead at the end it just calls a new is_builtin()
function to determine if the given item is a builtin function keyword, and
prunes from there. This paragraph is much longer than the actual solution.

Beyond that, there's just some extra warnings that I hadn't gotten around to
cleaning until this point, as well as some other odds and ends (comments, et
al).