Update basic.rasp, add some more "testing" files
basic.rasp is probably not going to include string manipulation methods, so
I've delegated those functions to go into a new library file called
"strings.rasp", which basic.rasp includes. Right now, it just holds the
formatting function :)

Along with strings.rasp, there's a new test file, factorial.rasp. This is just
for testing basic things that the compiler should be able to do right out of
the gate (like factorials!). Obviously this will change in the future, but I'm
glad to be able to test the compiler on something *real* and not just these
function definitions.