Add some include gather fixes, add start of bytecode and VM.
Include gatherer, for a time, wouldn't actually gather the functions and types
defined in the files it includes. This has been fixed, and works up to any
depth of included files. As I typed that, I need to check for include cycles.
Huge TODO.

The other thing that is added (it was just a lot easier to type `git add src/`)
is the addition of the *start* of bytecode generation. The stuff you see really
is just preliminary work-in-progress and will probably change a lot. I also
added a VM module for virtual machine related things, once I get bytecode more
polished up.