Add the start of bytecode and a VM, plus builtins.
Bytecode and VM are some of the more important parts to getting this thing off
the ground running, as they're the part of the program that actually *do*
things. Bytecode gets generated just fine, but I want to start looking more at
function parameters and verifying their types when those functions get called
(either a warning at compile-time or an error at run-time).

Another major addition is the creation of some builtins. Right now it's just
a few POSIX file manipulation functions (read, write, open, close). I want to
start working on some of the more important functions like if, match, car, cdr,
and the equality operators.

A final note for this commit is that there are a couple new dependencies: libc
and lazy_static, so we can have our C library functions as well as our static
hashmaps :)