Add varargs and optional parameters to function definitions.
This is much smaller commit than last time, but it's a good "commit" point. The
function gatherer now looks for optional arguments (anything after a `?'
character) and variable arguments (any list of parameters that ends with a
'...'). I haven't introduced any sort of "spread" operator, but I imagine the
varargs variable just gets encoded as a list at run-time.